requirements.txt
Flask==2.3.3
PyMuPDF==1.23.8
Werkzeug==2.3.7
google-genai==1.31.0
